factor out drawer widgets
