+++ packages/product-core/src/Session/index.ts
@@ -1,2 +1,3 @@
 export { default as ReferenceManagement } from './ReferenceManagement'
 export { default as Connections } from './Connections'
+export { default as DrawerWidgets } from './DrawerWidgets'
