--- products/jbrowse-web/src/sessionModel/Drawer.ts
@@ -1,62 +1,0 @@
-import { types } from 'mobx-state-tree'
-
-import PluginManager from '@jbrowse/core/PluginManager'
-import { localStorageGetItem } from '@jbrowse/core/util'
-
-const minDrawerWidth = 128
-
-export default function Drawer(pluginManager: PluginManager) {
-  return types
-    .model({
-      /**
-       * #property
-       */
-      drawerPosition: types.optional(
-        types.string,
-        () => localStorageGetItem('drawerPosition') || 'right',
-      ),
-      /**
-       * #property
-       */
-      drawerWidth: types.optional(
-        types.refinement(types.integer, width => width >= minDrawerWidth),
-        384,
-      ),
-    })
-    .actions(self => ({
-      /**
-       * #action
-       */
-      setDrawerPosition(arg: string) {
-        self.drawerPosition = arg
-      },
-
-      /**
-       * #action
-       */
-      updateDrawerWidth(drawerWidth: number) {
-        if (drawerWidth === self.drawerWidth) {
-          return self.drawerWidth
-        }
-        let newDrawerWidth = drawerWidth
-        if (newDrawerWidth < minDrawerWidth) {
-          newDrawerWidth = minDrawerWidth
-        }
-        self.drawerWidth = newDrawerWidth
-        return newDrawerWidth
-      },
-
-      /**
-       * #action
-       */
-      resizeDrawer(distance: number) {
-        if (self.drawerPosition === 'left') {
-          distance *= -1
-        }
-        const oldDrawerWidth = self.drawerWidth
-        const newDrawerWidth = this.updateDrawerWidth(oldDrawerWidth - distance)
-        const actualDistance = oldDrawerWidth - newDrawerWidth
-        return actualDistance
-      },
-    }))
-}
